+++ shared/java/InversionMode.java
@@ -2,6 +2,7 @@ package io.github.humbleui.skija;
 
 import org.jetbrains.annotations.*;
 
+// SkHighContrastFilter.h - SkHighContrastConfig::InvertStyle
 public enum InversionMode {
     NO,
     BRIGHTNESS,